third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/regress-532254-expected.txt

PASS Function("'use strict'; function f(eval,[x]){}") threw exception of type SyntaxError.
PASS !!Function("function f(eval,[x]){}") is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
